делает 5 задание
