+++ js/checks-length.js
@@ -1,0 +1,10 @@
+function checksLength ( str, length ) {
+  if (typeof str === 'string') {
+    return str.length <= length;
+  }
+  return NaN;
+}
+
+checksLength ();
+
+export { checksLength };
